Add "WordCombinator" token inbetween words/interpolations
When no whitespace occurs between word and/or interpolation tokens then
this must be obvious to the parser. This can be achieved by adding an
extra token inbetween when there's no whitespace.

This is now done using the CombinatorLoop, which peeks at the next input
value and adds this token when necessary.