Implement recursive value parsing
Add parseValues which recursively parses all values of a declaration, by
dividing it by functions, compound values, and values. It returns a
LinkedList (i.e. buffer) of these nodes which is then consumed inside a
BufferLoop.

It's not possible to divide this process up efficiently into multiple
steps of the parser stream.