Add extremely minified property prefixing
@rofrischmann has found a way to get the minimum substring for each
property that should be prefixed so that it's still unambiguous.

Now, each input property needs to be checked to see whether it starts
with one of those prefixes. Then if it does, those need to be checked
against what vendor-prefixes shall be added.

This is done by converting each char of each prefix into a bit on a
bitmap. Hence we'll end up with an array of bitmaps up to the maximum
length of all bitmaps.

A second bitmap is added that has a bit set when the last bit of a
prefix is reached and then a result is looked up which dictates whether
a vendor-prefix should be added.

The size benefit versus the whole list of properties that should be
prefixed is: +1.2kB - 0.4kB = +0.8kB gzip