Add support for partial interpolations
Partial interpolations i.e. mixins;

A heuristic is needed since the interpolations would otherwise
often be mistaken for the start of a selector including a selector
interpolation.

Instead the `mainLoop` (outside of selector/declaration parsing/buffering)
is now looking at tokens after interpolations. If the next token is on a
separate row (read: line) and is not a colon, then the interpolation
is regarded to be a partial.

The same goes for interpolations that are followed by a closing brace.